Delayed app object construction until the user clicks "save and run".
Changed how the step funciton is scheduled in the Tk event loop. Used a
combination of after_idle() with after() so that idle events get
processed - mainly the issue was that the alpha setting for the overlays
didn't have time to update so some of the overlays were hanging around.